+++ src/command-pallette/CommandPalletteUI/CommandPaletteSearchBar.scss
@@ -1,0 +1,7 @@
+.command-palette-search-bar {
+  width: var(--_ui5-v2-7-0_input_width) !important;
+}
+
+.ui5-shellbar-mid-content {
+  flex-grow: 1;
+}
