+++ src/command-pallette/CommandPalletteUI/CommandPaletteUI.scss
@@ -2,18 +2,15 @@
   position: fixed;
   width: 100%;
   height: 100%;
-
   display: flex;
   justify-content: center;
-
   z-index: 1000000;
-  background-color: rgba(0, 0, 0, 0.6);
 
   &__wrapper {
     width: 60vw;
     max-width: 700px;
     position: fixed;
-    margin-top: 20vh;
+    top: 0.5rem;
   }
 
   &__content {
@@ -23,5 +20,9 @@
     background: var(--sapList_Background);
     padding: 10px;
     border-radius: 10px;
+    box-shadow: 0 0 0.125rem 0
+        color-mix(in srgb, var(--sapContent_ShadowColor) 16%, transparent),
+      0 0.5rem 1rem 0
+        color-mix(in srgb, var(--sapContent_ShadowColor) 16%, transparent);
   }
 }
